+++ monkey/infection_monkey/payload/ransomware/file_selectors.py
@@ -1,4 +1,5 @@
 import filecmp
+import logging
 from pathlib import Path
 from typing import Iterable, Set
 
@@ -12,12 +13,29 @@ from infection_monkey.utils.file_utils import (
 
 from .consts import README_FILE_NAME, README_SRC
 
+logger = logging.getLogger(__name__)
+
 
 class ProductionSafeTargetFileSelector:
     def __init__(self, targeted_file_extensions: Set[str]):
         self._targeted_file_extensions = targeted_file_extensions
 
     def __call__(self, target_dir: Path) -> Iterable[Path]:
+        if not target_dir.exists():
+            logger.warning(f"Target directory {target_dir} does not exist")
+            return iter([])
+
+        if not target_dir.is_dir():
+            logger.warning(f"Target directory {target_dir} is not a directory")
+            return iter([])
+
+        if target_dir.is_symlink():
+            logger.warning(
+                "The ProductionSafeTargetFileSelector will not follow symlinks - skipping "
+                f"{target_dir}"
+            )
+            return iter([])
+
         file_filters = [
             file_extension_filter(self._targeted_file_extensions),
             is_not_shortcut_filter,